revert changes in SchemaResource
